Renaming all files to follow lowercase underscore
